Add HistoryProviderInitializeParameters
Sometimes, for many different reasons, we might need to add new parameters to existing methods with a large number of parameters. In this case, in order to fix a bug in SubscriptionDataReaderHistoryProvider we need to add the result handler to the Initialize method. Instead of adding yet another parameter, in this commit we introduce a new class which will be used as the only parameter to the Initialize method. This change allows us to add more parameters in the future without breaking the IHistoryProvider interface again and again.